verifier/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
httpx==0.27.2
python-jose[cryptography]==3.3.0
